refactor: boolean columns — refunds.restocked
refund_retail_save()'s `1 if restock else 0` -> plain `restock` (it
was already a Python bool from `f.get("restock") == "on"`), and
logic.consignment_balance()'s `r.restocked=1` -> `r.restocked=true`.
Updated the literal `::int` cast in generate_test_data.py's refunds
insert to a plain boolean expression. The other refunds INSERT
(service refunds, in app.py) never set restocked at all — it just
takes the column default — so it needed no change.

Verified with a direct insert against a fresh local Postgres 16
database: the resulting refunds row stored restocked as a real
`boolean` with the expected value.

+++ generate_test_data.py
@@ -381,7 +381,7 @@ def main():
         SELECT rtype,
                (current_date - (floor(random()*{HISTORY_MONTHS}*30))::int),
                round((5000+random()*50000)::numeric,0)::float,
-               (rtype='retail' AND random()<0.5)::int,
+               (rtype='retail' AND random()<0.5),
                CASE WHEN rtype='service' AND random()<0.7 THEN vid[1 + floor(random()*array_length(vid,1))::int] ELSE NULL END,
                NULL, NULL,
                uid[1 + floor(random()*array_length(uid,1))::int], now()::text
